Fix interface conversion

+++ discord/currency.go
@@ -24,7 +24,7 @@ func (b *Bot) currencyBasics(s *discordgo.Session, m *discordgo.MessageCreate) {
 		if !exists {
 			user.Metadata["lastdaily"] = time.Now().Unix()
 		} else {
-			diff := time.Now().Unix() - user.Metadata["lastdaily"].(int64)
+			diff := time.Now().Unix() - int64(user.Metadata["lastdaily"].(float64))
 			if (diff) < 86400 { // less than a day
 				s.ChannelMessageSend(m.ChannelID, "You still need to wait "+strconv.Itoa(int(diff/3600))+" hours.")
 				return
